apps/consumer/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.3
python-dateutil==2.8.2